+++ src/components/hall-components/menu-section/menu-section.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react'
-import MenuItens from '../menu-itens/menu-itens'
+import MenuItens from '../hall/menu-items/menu-items'
 import './menu-section.css'
 
 export default function MenuSection({
