+++ src/index.js
@@ -1,9 +1,7 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 import { Routes } from './routes'
-
-
+import React from 'react'
+import ReactDOM from 'react-dom'
 
 ReactDOM.render(
   <BrowserRouter>
