+++ src/pages/Login/login.js
@@ -2,8 +2,8 @@ import { Link, useHistory } from 'react-router-dom'
 import { Fragment, useState } from 'react'
 import { authUser } from '../../services/index'
 import Button from '../../components/Button/Button'
-import Input from '../../components/Input/Input'
-import Footer from '../../components/Footer/Footer'
+import Input from '../../components/generic-components/input/input'
+import Footer from '../../components/generic-components/footer/footer'
 import logo from '../../assets/logo.png'
 import './login.css'
