+++ src/pages/Signup/signup.js
@@ -1,9 +1,9 @@
 import { Fragment, useState } from 'react'
-import { createUser } from '../../services/index'
-import { Link, useHistory } from 'react-router-dom';
-import Button from '../../components/Button/Button'
-import Input from '../../components/Input/Input.js'
-import Footer from '../../components/Footer/Footer'
+import { Link, useHistory } from 'react-router-dom'
+import { createUser } from '../../services/services'
+import Button from '../../components/button/button'
+import Input from '../../components/generic-components/input/input'
+import Footer from '../../components/generic-components/footer/footer'
 import logo from '../../assets/logo.png'
 import './signup.css'
