--- src/pages/StatusOrder/index.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { Fragment } from 'react'
-import Navbar from '../../components/Navbar/Navbar'
-export const StatusOrder = () => {
-
-
-  return (
-    <Fragment>
-      <header>
-        <Navbar />
-      </header>
-    </Fragment >
-  )
-}
