+++ src/pages/status-order/status-order.js
@@ -1,0 +1,14 @@
+import { Fragment } from 'react'
+import React from 'react'
+import Navbar from '../../components/generic-components/navbar/navbar'
+import './status-order.css'
+
+export const StatusOrder = () => {
+  return (
+    <Fragment>
+      <header>
+        <Navbar />
+      </header>
+    </Fragment >
+  )
+}
